requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
gunicorn==19.6.0
psycopg2==2.6.2
requests==2.11.1
static3==0.7.0
wheel==0.24.0
PyYAML==3.11
